Fix: New Template blood_pressure
