+++ src/main/java/com/myobservation/auth/dto/UserResponse.java
@@ -4,7 +4,6 @@ import java.util.Set;
 
 /**
  * Representa la respuesta de datos de un usuario.
- *
  * Esta clase se usa para enviar información de usuario desde la API sin exponer datos sensibles como la contraseña.
  */
 public class UserResponse {
