+++ src/main/java/com/myobservation/ehr/pojos/BloodPressureCompositionContainment.java
@@ -1,0 +1,62 @@
+package com.myobservation.ehr.pojos;
+
+import com.myobservation.ehr.pojos.definition.BloodPressureObservation;
+import com.nedap.archie.rm.archetyped.FeederAudit;
+import com.nedap.archie.rm.datastructures.Cluster;
+import com.nedap.archie.rm.generic.Participation;
+import com.nedap.archie.rm.generic.PartyIdentified;
+import com.nedap.archie.rm.generic.PartyProxy;
+import java.lang.String;
+import java.time.temporal.TemporalAccessor;
+import org.ehrbase.openehr.sdk.generator.commons.aql.containment.Containment;
+import org.ehrbase.openehr.sdk.generator.commons.aql.field.AqlFieldImp;
+import org.ehrbase.openehr.sdk.generator.commons.aql.field.ListAqlFieldImp;
+import org.ehrbase.openehr.sdk.generator.commons.aql.field.ListSelectAqlField;
+import org.ehrbase.openehr.sdk.generator.commons.aql.field.SelectAqlField;
+import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.Category;
+import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.Language;
+import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.NullFlavour;
+import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.Setting;
+import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.Territory;
+
+public class BloodPressureCompositionContainment extends Containment {
+  public SelectAqlField<BloodPressureComposition> BLOOD_PRESSURE_COMPOSITION = new AqlFieldImp<BloodPressureComposition>(BloodPressureComposition.class, "", "BloodPressureComposition", BloodPressureComposition.class, this);
+
+  public SelectAqlField<Category> CATEGORY_DEFINING_CODE = new AqlFieldImp<Category>(BloodPressureComposition.class, "/category|defining_code", "categoryDefiningCode", Category.class, this);
+
+  public SelectAqlField<String> REPORT_ID_VALUE = new AqlFieldImp<String>(BloodPressureComposition.class, "/context/other_context[at0001]/items[at0002]/value|value", "reportIdValue", String.class, this);
+
+  public SelectAqlField<NullFlavour> REPORT_ID_NULL_FLAVOUR_DEFINING_CODE = new AqlFieldImp<NullFlavour>(BloodPressureComposition.class, "/context/other_context[at0001]/items[at0002]/null_flavour|defining_code", "reportIdNullFlavourDefiningCode", NullFlavour.class, this);
+
+  public ListSelectAqlField<Cluster> EXTENSION = new ListAqlFieldImp<Cluster>(BloodPressureComposition.class, "/context/other_context[at0001]/items[at0006]", "extension", Cluster.class, this);
+
+  public SelectAqlField<TemporalAccessor> START_TIME_VALUE = new AqlFieldImp<TemporalAccessor>(BloodPressureComposition.class, "/context/start_time|value", "startTimeValue", TemporalAccessor.class, this);
+
+  public ListSelectAqlField<Participation> PARTICIPATIONS = new ListAqlFieldImp<Participation>(BloodPressureComposition.class, "/context/participations", "participations", Participation.class, this);
+
+  public SelectAqlField<TemporalAccessor> END_TIME_VALUE = new AqlFieldImp<TemporalAccessor>(BloodPressureComposition.class, "/context/end_time|value", "endTimeValue", TemporalAccessor.class, this);
+
+  public SelectAqlField<String> LOCATION = new AqlFieldImp<String>(BloodPressureComposition.class, "/context/location", "location", String.class, this);
+
+  public SelectAqlField<PartyIdentified> HEALTH_CARE_FACILITY = new AqlFieldImp<PartyIdentified>(BloodPressureComposition.class, "/context/health_care_facility", "healthCareFacility", PartyIdentified.class, this);
+
+  public SelectAqlField<Setting> SETTING_DEFINING_CODE = new AqlFieldImp<Setting>(BloodPressureComposition.class, "/context/setting|defining_code", "settingDefiningCode", Setting.class, this);
+
+  public SelectAqlField<BloodPressureObservation> BLOOD_PRESSURE = new AqlFieldImp<BloodPressureObservation>(BloodPressureComposition.class, "/content[openEHR-EHR-OBSERVATION.blood_pressure.v2]", "bloodPressure", BloodPressureObservation.class, this);
+
+  public SelectAqlField<PartyProxy> COMPOSER = new AqlFieldImp<PartyProxy>(BloodPressureComposition.class, "/composer", "composer", PartyProxy.class, this);
+
+  public SelectAqlField<Language> LANGUAGE = new AqlFieldImp<Language>(BloodPressureComposition.class, "/language", "language", Language.class, this);
+
+  public SelectAqlField<FeederAudit> FEEDER_AUDIT = new AqlFieldImp<FeederAudit>(BloodPressureComposition.class, "/feeder_audit", "feederAudit", FeederAudit.class, this);
+
+  public SelectAqlField<Territory> TERRITORY = new AqlFieldImp<Territory>(BloodPressureComposition.class, "/territory", "territory", Territory.class, this);
+
+  private BloodPressureCompositionContainment() {
+    super("openEHR-EHR-COMPOSITION.report.v1");
+  }
+
+  public static BloodPressureCompositionContainment getInstance() {
+    return new BloodPressureCompositionContainment();
+  }
+}
