+++ src/main/java/com/myobservation/ehr/pojos/definition/BloodPressureLocationOfMeasurementChoice.java
@@ -1,12 +1,11 @@
-package com.myobservation.ehr.definition;
+package com.myobservation.ehr.pojos.definition;
 
 import javax.annotation.processing.Generated;
 
 @Generated(
     value = "org.ehrbase.openehr.sdk.generator.ClassGenerator",
-    date = "2025-05-18T10:02:55.944869200+02:00",
+    date = "2025-05-22T10:35:38.140521+02:00",
     comments = "https://github.com/ehrbase/openEHR_SDK Version: 2.23.0-SNAPSHOT"
 )
 public interface BloodPressureLocationOfMeasurementChoice {
-
 }
