+++ src/main/java/com/myobservation/ehr/pojos/definition/BloodPressureLocationOfMeasurementDvText.java
@@ -1,4 +1,4 @@
-package com.myobservation.ehr.definition;
+package com.myobservation.ehr.pojos.definition;
 
 import java.lang.String;
 import javax.annotation.processing.Generated;
@@ -10,13 +10,13 @@ import org.ehrbase.openehr.sdk.generator.commons.interfaces.RMEntity;
 @Entity
 @Generated(
     value = "org.ehrbase.openehr.sdk.generator.ClassGenerator",
-    date = "2025-05-18T10:02:55.944869200+02:00",
+    date = "2025-05-22T10:35:38.140521+02:00",
     comments = "https://github.com/ehrbase/openEHR_SDK Version: 2.23.0-SNAPSHOT"
 )
 @OptionFor("DV_TEXT")
 public class BloodPressureLocationOfMeasurementDvText implements RMEntity, BloodPressureLocationOfMeasurementChoice {
   /**
-   * Path: Presión Sanguínea/Blood pressure/Location of measurement/Location of measurement
+   * Path: blood_pressure/Blood pressure/Location of measurement/Location of measurement
    * Description: Simple body site where blood pressure was measured.
    */
   @Path("|value")
