+++ src/main/java/com/myobservation/ehr/pojos/definition/BloodPressureObservation.java
@@ -1,4 +1,4 @@
-package com.myobservation.ehr.definition;
+package com.myobservation.ehr.pojos.definition;
 
 import com.nedap.archie.rm.archetyped.FeederAudit;
 import com.nedap.archie.rm.datastructures.Cluster;
@@ -18,51 +18,51 @@ import org.ehrbase.openehr.sdk.generator.commons.shareddefinition.NullFlavour;
 @Archetype("openEHR-EHR-OBSERVATION.blood_pressure.v2")
 @Generated(
     value = "org.ehrbase.openehr.sdk.generator.ClassGenerator",
-    date = "2025-05-18T10:02:55.882871+02:00",
+    date = "2025-05-22T10:35:38.100095200+02:00",
     comments = "https://github.com/ehrbase/openEHR_SDK Version: 2.23.0-SNAPSHOT"
 )
 public class BloodPressureObservation implements EntryEntity {
   /**
-   * Path: Presión Sanguínea/Blood pressure/origin
+   * Path: blood_pressure/Blood pressure/origin
    */
   @Path("/data[at0001]/origin|value")
   private TemporalAccessor originValue;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Tree/Location of measurement/null_flavour
+   * Path: blood_pressure/Blood pressure/Tree/Location of measurement/null_flavour
    */
   @Path("/protocol[at0011]/items[at0014]/null_flavour|defining_code")
   private NullFlavour locationOfMeasurementNullFlavourDefiningCode;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Structured measurement location
+   * Path: blood_pressure/Blood pressure/Structured measurement location
    * Description: Structured anatomical location of where the measurement was taken.
    */
   @Path("/protocol[at0011]/items[at1057]")
   private List<Cluster> structuredMeasurementLocation;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Method
+   * Path: blood_pressure/Blood pressure/Method
    * Description: Method of measurement of blood pressure.
    */
   @Path("/protocol[at0011]/items[at1035]/value|defining_code")
   private MethodDefiningCode methodDefiningCode;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Tree/Method/null_flavour
+   * Path: blood_pressure/Blood pressure/Tree/Method/null_flavour
    */
   @Path("/protocol[at0011]/items[at1035]/null_flavour|defining_code")
   private NullFlavour methodNullFlavourDefiningCode;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Device
+   * Path: blood_pressure/Blood pressure/Device
    * Description: Details about sphygmomanometer or other device used to measure the blood pressure.
    */
   @Path("/protocol[at0011]/items[at1025]")
   private Cluster device;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Extension
+   * Path: blood_pressure/Blood pressure/Extension
    * Description: Additional information required to capture local context or to align with other reference models/formalisms.
    * Comment: For example: Local hospital departmental infomation or additional metadata to align with FHIR or CIMI equivalents.
    */
@@ -70,25 +70,25 @@ public class BloodPressureObservation implements EntryEntity {
   private List<Cluster> extension;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/subject
+   * Path: blood_pressure/Blood pressure/subject
    */
   @Path("/subject")
   private PartyProxy subject;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/language
+   * Path: blood_pressure/Blood pressure/language
    */
   @Path("/language")
   private Language language;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/feeder_audit
+   * Path: blood_pressure/Blood pressure/feeder_audit
    */
   @Path("/feeder_audit")
   private FeederAudit feederAudit;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Any event
+   * Path: blood_pressure/Blood pressure/Any event
    * Description: Default, unspecified point in time or interval event which may be explicitly defined in a template or at run-time.
    */
   @Path("/data[at0001]/events[at0006]")
@@ -96,7 +96,7 @@ public class BloodPressureObservation implements EntryEntity {
   private List<BloodPressureAnyEventChoice> anyEvent;
 
   /**
-   * Path: Presión Sanguínea/Blood pressure/Location of measurement
+   * Path: blood_pressure/Blood pressure/Location of measurement
    * Description: Simple body site where blood pressure was measured.
    */
   @Path("/protocol[at0011]/items[at0014]/value")
