+++ src/main/resources/application.properties
@@ -36,6 +36,7 @@ ehrbase.password=ehrbase_restricted
 ehrbase.templates.path=src/main/resources/templates
 logging.level.org.ehrbase.openehr.sdk=DEBUG
 logging.level.org.springframework.web.client.RestTemplate=DEBUG
+
 # Logging
 logging.level.org.springframework=INFO
 logging.level.com.myobservation=DEBUG
